Create SatelliteDriver.java
Here is the basic GUI for the program.  It needs work.  It could use a reset button to clear the data and start with new satellite data.  Also the action listeners need to be implemented.  The class will need instantiate a new Satellite object.  Do you think the main method should be a separate file?